Dependency Injection y EF Core. Se agrega configuracion del contexto con Fluent Api. Insertar una nuevo producto Se agregan validaciones con FluentValidations. Se agrega actualización y eliminación de Posts. Se implementa el Repository Pattern y Unit of Work Se agrega manejo de excepción global. Se agrega filtro de datos. Paginación inicial. Paginación configurable.

+++ VamTech.Ecommerce.Core/Interfaces/IRepository.cs
@@ -9,9 +9,9 @@ namespace VamTech.Ecommerce.Core.Interfaces
     public interface IRepository<T> where T : BaseEntity
     {
         IEnumerable<T> GetAll();
-        Task<T> GetById(int id);
+        Task<T> GetById(long id);
         Task Add(T entity);
         void Update(T entity);
-        Task Delete(int id);
+        Task Delete(long id);
     }
 }
